requirements-ingestion.txt
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
defusedxml>=0.7
